+++ StartUp/BusinessLogic/UserService.cs
@@ -14,29 +14,22 @@ namespace BusinessLogic
     public class UserService : IUserService
     {
         private readonly IRepository<User> _userRepository;
+        private readonly IRepository<Role> _roleRepository;
+        private readonly IRepository<Pharmacy> _pharmacyRepository;
+        private readonly IRepository<Invitation> _invitationRepository;
 
-        public UserService(IRepository<User> userRepository)
+        public UserService(IRepository<User> userRepository, IRepository<Role> roleRepository, 
+            IRepository<Pharmacy> pharmacyRepository, IRepository<Invitation> invitationRepository)
         {
             _userRepository = userRepository;
+            _roleRepository = roleRepository;
+            _pharmacyRepository = pharmacyRepository;
+            _invitationRepository = invitationRepository;
         }
 
         public List<User> GetAllUser(UserSearchCriteria searchCriteria)
         {
-            var emailCriteria = searchCriteria.Email?.ToLower() ?? string.Empty;
-            var passwordCriteria = searchCriteria.Password?.ToLower() ?? string.Empty;
-            var addressCriteria = searchCriteria.Address?.ToLower() ?? string.Empty;
-            var invitationCriteria = searchCriteria.Invitation ?? null;
-            var registerDateCriteria = searchCriteria.RegisterDate?.ToString() ?? string.Empty;
-            var pharmacyCriteria = searchCriteria.Pharmacy ?? null;
-            var rolCriteria = searchCriteria.Roles ?? null;
-
-            Expression<Func<User, bool>> userFilter = user =>
-                user.Email.ToLower().Contains(emailCriteria) &&
-                user.Password.ToLower().Contains(passwordCriteria) &&
-                user.Address.ToLower().Contains(addressCriteria) &&
-                user.Invitation == invitationCriteria &&
-                user.RegisterDate.ToString().Contains(registerDateCriteria) &&
-                user.Pharmacy == pharmacyCriteria && user.Roles == rolCriteria;
+            Expression<Func<User, bool>> userFilter = user => true;
 
             return _userRepository.GetAllByExpression(userFilter).ToList();
         }
@@ -48,8 +41,26 @@ namespace BusinessLogic
 
             User UserSaved = _userRepository.GetOneByExpression(u => u.Id == userId);
 
-            validator.ValidateUserNotNull(UserSaved, $"The user {userId} not exist");
+            if(UserSaved == null)
+            {
+                throw new InputException($"The user {userId} not exist");
+            }
 
+            return UserSaved;
+        }
+        
+        public User GetSpecificUserByUserName(string username)
+        {
+            Validator validator = new Validator();
+            validator.ValidateString(username.ToString(), "User empty");
+
+            User UserSaved = _userRepository.GetOneByExpression(u => u.Invitation.UserName == username);
+
+            if (UserSaved == null)
+            {
+                throw new InputException($"The user {username} not exist");
+            }
+            
             return UserSaved;
         }
 
@@ -57,10 +68,20 @@ namespace BusinessLogic
         {
             user.IsValidUser();
             EmailNotExistInDataBase(user);
-            user.VerifyInvitationStateIsAvailable();
+            user.Invitation = _invitationRepository.GetOneByExpression(i => i.UserName == user.Invitation.UserName);
+            user.VerifyInvitationExist();
+            user.VerifyInvitationState();
+            user.VerifyInvitationRoles();
+            user.VerifyInvitationPharmacy();
+            user.VerifyRolesAndPharmacy();
             user.ChangeStatusInvitation();
             user.RegisterDate = DateTime.Now;
 
+            if ((user.Pharmacy != null) && !user.Roles.Permission.Contains("administrator"))
+            {
+                user.Pharmacy = _pharmacyRepository.GetOneByExpression(p => p.Name == user.Pharmacy.Name);
+            }
+            user.Roles = _roleRepository.GetOneByExpression(r=>r.Permission == user.Roles.Permission);
             _userRepository.InsertOne(user);
             _userRepository.Save();
 
@@ -69,24 +90,33 @@ namespace BusinessLogic
 
         public User UpdateUser(int userId, User updateduser)
         {
-            Validator validator = new Validator();
-            validator.ValidateUserNotNull(updateduser, "User empty");
-            validator.ValidateString(userId.ToString(), "UserId empty");
-
             updateduser.IsValidUser();
-
+            
             var userStored = GetSpecificUser(userId);
-            validator.ValidateUserNotNull(userStored, "User not exist");
+            if(userStored == null)
+            {
+                throw new InputException("User not exist");
+            }
+
+            updateduser.Invitation = _invitationRepository.GetOneByExpression(i => i.UserName == updateduser.Invitation.UserName);
+            updateduser.VerifyInvitationExist();
+            updateduser.VerifyInvitationState();
+            updateduser.VerifyInvitationRoles();
+            updateduser.VerifyInvitationPharmacy();
+            updateduser.VerifyRolesAndPharmacy();
 
             userStored.Email = updateduser.Email;
-            userStored.Password = updateduser.Password;
             userStored.Address = updateduser.Address;
             userStored.RegisterDate = updateduser.RegisterDate;
             userStored.Password = updateduser.Password;
             userStored.Invitation = updateduser.Invitation;
-            userStored.Pharmacy = updateduser.Pharmacy;
-            userStored.Roles = updateduser.Roles;
+            userStored.Roles = _roleRepository.GetOneByExpression(r => r.Permission == updateduser.Roles.Permission);
 
+            if (updateduser.Pharmacy != null)
+            {
+                userStored.Pharmacy = _pharmacyRepository.GetOneByExpression(p => p.Name == updateduser.Pharmacy.Name);
+            }
+            
             _userRepository.UpdateOne(userStored);
             _userRepository.Save();
 
@@ -97,8 +127,12 @@ namespace BusinessLogic
         {
             Validator validator = new Validator();
             validator.ValidateString(userId.ToString(), "UserID empty");
+
             var userStored = GetSpecificUser(userId);
-            validator.ValidateUserNotNull(userStored, "User not exist");
+            if (userStored == null)
+            {
+                throw new InputException("User not exist");
+            }
 
             _userRepository.DeleteOne(userStored);
             _userRepository.Save();
@@ -106,9 +140,11 @@ namespace BusinessLogic
 
         private void EmailNotExistInDataBase(User user)
         {
-            Validator validator = new Validator();
-            validator.ValidateUserNotNull(user, "User empty");
-
+            if (user == null)
+            {
+                throw new InputException("User empty");
+            }
+            
             User userSaved = _userRepository.GetOneByExpression(u => u.Email == user.Email);
 
             if (userSaved != null)
@@ -119,14 +155,10 @@ namespace BusinessLogic
 
         public void SaveToken(User user, string token)
         {
-            Validator validator = new Validator();
-            validator.ValidateString(token, "Token empty");
-
-            var userSalved = _userRepository.GetOneByExpression(u => u.Id == user.Id);
-
-            validator.ValidateUserNull((userSalved), "User not exist in database");
-
+            var userSalved = _userRepository.GetOneByExpression(u => u.Invitation.UserName == user.Invitation.UserName);
+            
             userSalved.Token = token;
+            _userRepository.UpdateOne(userSalved);
             _userRepository.Save();
         }
